Use map.kv_entry in map literal (#122)
I've been working on improving support for map literals, and one of
challenging things about that is getting the list of key-value pairs
passed to the `map.literal` function to properly collapse down to a
value. The list automation for this only works if each element of the
list can be evaluated to a value, which isn't the case for these
elements.

I experimented in perennial with allowing tuples to be directly
evaluated into a `PairV`, but that causes a host of other issues with
struct fields, which are also modeled as a pair (field name to value).
Eventually @tchajed and I decided to introduce a new type of value,
`kv_entry` specifically for map literals so that the proof automation
could be developed without interfering with structs or any other place
were pairs of values appear in perennial.

This PR adjusts goose to output map literals with the `kv_entry` wrapper
and provide a unit test which will have a corresponding proof in
perennial (PR imminent).