Better support generics with known underlying type (#125)
If a generic function has an input `T ~[]E` and then an input `v T`,
slice operations should be available on `v`, and if the element type is
needed, it is known to be `E`. This commit improves that support,
including some (but probably not all) cases of the same issue for maps
and channels.

The upshot of this is that we can now translate
[slices.Clone](https://cs.opensource.google/go/go/+/refs/tags/go1.25.0:src/slices/slices.go;l=353).

Fixes #117.