Keep go_type opaque;
need to fix handling for generics it's the proofgen `Foo.ty` that should
be made Opaque to avoid TC search from picking the "wrong" `Type`. The
GooseLang function to construct the val representation of the type
should *also* be kept Opaque for this purpose to avoid the wrong
`wp_type_Foo` from being applied. This commit handles opaqueness for the
GooseLang function